+++ gfx_utils/matrix.h
@@ -22,7 +22,7 @@ void scalar_multiply(matrix *m, int a);
 matrix *identity_matrix(int n);
 
 matrix *get_translation_matrix(int dx, int dy, int dz);
-matrix *get_scale_matrix(int x_fac, int y_fac, int z_fac);
+matrix *get_scale_matrix(double x_fac, double y_fac, double z_fac);
 matrix *get_rot_x_matrix(double theta);
 matrix *get_rot_y_matrix(double theta);
 matrix *get_rot_z_matrix(double theta);
